refactor: moved getProviders to auth-options
